verif/template_netlist/requirements.txt
z3-solver==4.8.15 # later versions produce ~100x performance degradation (MR#1126)
pyyaml
ruamel.yaml
# click
